Used dtype_fallback, moved dtype_fallback from onnx.py to device.py
