+++ tinygrad/gradient.py
@@ -3,14 +3,13 @@ import math, dataclasses
 from tinygrad.uop.ops import UOp, PatternMatcher, UPat, Ops, all_metadata
 from tinygrad.helpers import argsort
 from tinygrad.dtype import dtypes
-from tinygrad.device import Device, is_dtype_supported
+from tinygrad.device import dtype_fallback
 
 def reduce_gradient(ctx:UOp, ret:UOp):
   def to_inp_shape(x): return x.reshape(x.shape+(1,)*(len(ret.src[0].shape)-len(x.shape))).expand(ret.src[0].shape)
   if ret.arg[0] == Ops.ADD: return (to_inp_shape(ctx),)
   if ret.arg[0] == Ops.MAX:
-    dtype = dtypes.float64 if is_dtype_supported(dtypes.float64, Device.DEFAULT) else dtypes.float32  # METAL does not support float64
-    max_is_1s = ret.src[0].eq(to_inp_shape(ret)).cast(dtype)
+    max_is_1s = ret.src[0].eq(to_inp_shape(ret)).cast(dtype_fallback(dtypes.float64, "reduce_gradient")) # METAL does not support float64
     div = to_inp_shape(max_is_1s.r(Ops.ADD, ret.arg[1]))
     return ((max_is_1s/div).cast(ctx.dtype) * to_inp_shape(ctx),)
   if ret.arg[0] == Ops.MUL: return (to_inp_shape(ctx * ret) / ret.src[0],)
